requirements.txt
tensorflow==2.15.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
pillow==10.1.0
